Simplified widget init code.
Turns out I don't really need to reset the model - it seems that the issues the original implementations had (which were fixed in the previous commits) prevented the view to properly update the model.

Not sure if I should be sad or happy that all those hacks I thought up to make it work, ended up being, well, just hacks.