Changed how the views are initially notified.
This is also a better approach how to set treeView column sizes.

The issue now is how to add custom styling to specific column, and it seems that we would be better off implementing the tree model via QStandardItemModel, instead of QAbstractItemModel.

QStandardItemModel also has the functionality required to get item indexes, etc, which we had to implement ourselves in QAbstractItemModel.
Will rework to it & see how it fares performance wise.